# apis/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Moadim Server API",
    "description": "REST API for managing routines",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.5.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "This server"
    }
  ],
  "paths": {
...
    "/routines": {
      "get": {
        "tags": [
          "crate::routes::list_routines"
        ],
        "summary": "`GET /routines` — list routines, optionally filtered and sorted by repository.",
        "operationId": "list_routines",
        "parameters": [
          {
            "name": "repository",
            "in": "query",
            "description": "Keep only routines with at least one repository whose URL contains this\nsubstring (case-insensitive). Empty or absent keeps every routine.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Field to sort by (default: creation time).",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RoutineSort"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Sort direction (default: ascending).",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder"
            }
          },
          {
            "name": "local_only",
            "in": "query",
            "description": "When `true`, only return routines whose `machines` list includes the current machine.\nDefaults to `false` (return all routines, preserving backwards compatibility).",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "include_prompts",
            "in": "query",
            "description": "When `true`, include each routine's `prompt` in the response. Defaults to `false`:\nthe prompt (often the largest field) is omitted so listings stay compact. Fetch a\nsingle routine with `svc_get` / `GET /routines/{id}` to always see its prompt.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoutineResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "crate::routes::create_routine"
        ],
        "summary": "`POST /routines` — create a new routine.",
        "operationId": "create_routine",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoutineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid cron expression"
          }
        }
      }
    },
...
    "/routines/runs": {
      "get": {
        "tags": [
          "crate::routines"
        ],
        "summary": "`GET /routines/runs` — the most recent runs across every routine, newest first. Backs the\noverview \"recent runs\" panel with a single workbench scan instead of one request per routine.",
        "operationId": "get_all_runs",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Cap on the number of runs returned (default: `DEFAULT_FLEET_RUNS_LIMIT`).",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FleetRunSummary"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/routines/{id}": {
      "get": {
        "tags": [
          "crate::routes::get_routine"
        ],
        "summary": "`GET /routines/{id}` — retrieve a single routine by UUID.",
        "operationId": "get_routine",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Routine UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "tags": [
          "crate::routes::update_routine"
        ],
        "summary": "`PUT /routines/{id}` — alias for `PATCH`: a partial-merge update, not a full replace.",
        "description": "Fields omitted from the body are retained from the existing record, exactly as with `PATCH`.\nA client expecting RFC 7231 full-resource-replacement semantics (omitted fields reset to\ndefault) should not rely on this route for that; use `PATCH` and set every field explicitly.",
        "operationId": "replace",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Routine UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoutineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "tags": [
          "crate::routes::delete_routine"
        ],
        "summary": "`DELETE /routines/{id}` — delete a routine by UUID.",
        "operationId": "delete_routine",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Routine UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "patch": {
        "tags": [
          "crate::routes::update_routine"
        ],
        "summary": "`PATCH /routines/{id}` — partially update a routine.",
        "operationId": "update_routine",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Routine UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoutineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/routines/{id}/flags": {
      "get": {
        "tags": [
          "crate::routines"
        ],
        "summary": "`GET /routines/{id}/flags` — list open flags raised against a routine.",
        "operationId": "list_flags",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Routine UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Flag"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "post": {
        "tags": [
          "crate::routines"
        ],
        "summary": "`POST /routines/{id}/flags` — raise a new flag against a routine.",
        "operationId": "create_flag",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Routine UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFlagRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Flag"
                }
              }
            }
          },
          "400": {
            "description": "Invalid type/description/scope"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/routines/{id}/flags/{filename}": {
      "delete": {
        "tags": [
          "crate::routines"
        ],
        "summary": "`DELETE /routines/{id}/flags/{filename}` — resolve (delete) a flag.",
        "operationId": "resolve_flag",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Routine UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filename",
            "in": "path",
            "description": "Flag filename, as returned by create/list",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Resolved"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/routines/{id}/logs": {
      "get": {
        "tags": [
          "crate::routines"
        ],
        "summary": "`GET /routines/{id}/logs` — return the newest workbench `agent.log` as plain text.",
        "operationId": "get_logs",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Routine UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Log file contents as plain text"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/routines/{id}/prompt-preview": {
      "get": {
        "tags": [
          "crate::routines"
        ],
        "summary": "`GET /routines/{id}/prompt-preview` — the exact prompt body a run would receive, computed\nin-memory with no workbench, `prompt.md` write, or agent launch (issue #391). Does not include\nthe routine-origin disclosure written separately to `CLAUDE.md` at trigger time.",
        "operationId": "get_prompt_preview",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Routine UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Composed prompt body as plain text"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/routines/{id}/runs": {
      "get": {
        "tags": [
          "crate::routes::list_routine_runs"
        ],
        "summary": "`GET /routines/{id}/runs` — list every run workbench for the routine, newest first.",
        "operationId": "list_routine_runs",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Routine UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunSummary"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/routines/{id}/scheduled-trigger": {
      "post": {
        "tags": [
          "crate::routines"
        ],
        "summary": "`POST /routines/{id}/scheduled-trigger` — run a routine on its schedule.",
        "description": "The daemon-side endpoint the generated crontab line invokes (`moadim schedule trigger <id>`).\nUnlike [`crate::routes::trigger_routine::trigger_routine`] it does not record a manual\ntrigger; the spawned command records the scheduled timestamp itself. See\n[`svc_trigger_scheduled`].",
        "operationId": "scheduled_trigger",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Routine UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Routine"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/routines/{id}/trigger": {
      "post": {
        "tags": [
          "crate::routes::trigger_routine"
        ],
        "summary": "`POST /routines/{id}/trigger` — manually trigger a routine outside its schedule.",
        "description": "Refuses (423, distinct message) when the routine is disabled or in power-saving mode. See\n[`crate::routines::svc_trigger`].",
        "operationId": "trigger_routine",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Routine UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Routine"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RoutineResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Routine",
            "description": "The underlying routine."
          },
          {
            "type": "object",
            "required": [
              "agent_registered",
              "agent_command_available",
              "file_path",
              "flag_count",
              "is_running",
              "env_keys"
            ],
            "properties": {
              "agent_command_available": {
                "type": "boolean",
                "description": "`true` if the agent config's `command` (e.g. `claude`, `codex`) resolves to an executable\non the daemon's `PATH`. Distinct from [`Self::agent_registered`]: a routine can have a\npresent, well-formed agent config yet reference a binary that isn't installed, in which\ncase the cron firing launches a tmux session that dies immediately with \"command not\nfound\" — a silent no-op indistinguishable from a healthy routine by `agent_registered`\nalone. `false` whenever the agent config is missing, unreadable, or malformed, since no\n`command` can be resolved in that case either."
              },
              "agent_registered": {
                "type": "boolean",
                "description": "`true` if an agent config exists at `~/.config/moadim/agents/<agent>.toml` *and* parses\nsuccessfully. A present-but-malformed config is silently dropped at crontab-sync time, so\nit reports `false` here too — file existence alone is not \"registered\"."
              },
              "env_keys": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "Names (never values) of every environment variable set for this routine, merging the\ntracked `routine.toml` `[env]` table with the untracked `routine.local.toml` sidecar\n(secrets), deduplicated and sorted. Lets a client show *what* is configured without ever\nexposing a secret value over the API (issue #408)."
              },
              "file_path": {
                "type": "string",
                "description": "Absolute path to the routine's `routine.toml` file on disk."
              },
              "flag_count": {
                "type": "integer",
                "description": "Number of open flags raised against this routine (see [`super::flags`]). Surfaced here so\nlistings can badge it without a separate `list_flags` round-trip per routine.",
                "minimum": 0
              },
              "is_running": {
                "type": "boolean",
                "description": "`true` if any fire of this routine currently has a live tmux session — i.e. an agent is\nrunning right now. Derived by probing for a session under the routine's\n`moadim-{slug}-` prefix (the same overlap-guard check `svc_trigger` uses, #514), not\npersisted. `false` whenever no `tmux` binary is available, mirroring the probe's existing\nbest-effort \"no tmux, nothing running\" stance. See issue #438."
              },
              "next_run_at": {
                "type": [
                  "integer",
                  "null"
                ],
                "format": "int64",
                "description": "Unix epoch seconds of this routine's next scheduled fire, in the host's local timezone\n(matching crontab semantics) — the future counterpart to `last_scheduled_trigger_at`.\n`None` when disabled, globally locked, or `schedule` is unparseable or has no upcoming\nfire (e.g. `@reboot`). See issue #369.",
                "minimum": 0
              },
              "schedule_description": {
                "type": [
                  "string",
                  "null"
                ],
                "description": "Human-readable description of the schedule, including the timezone the\ncron expression is interpreted in, or `null` if it cannot be parsed."
              },
              "timezone": {
                "type": [
                  "string",
                  "null"
                ],
                "description": "IANA name of the local timezone the schedule is interpreted in (e.g.\n`\"Asia/Jerusalem\"`), or `null` if it cannot be determined. Cron\nexpressions are evaluated in this timezone, **not** UTC."
              }
            }
          }
        ],
        "description": "A [`Routine`] enriched with derived, non-persisted fields for API responses."
      },
      "RoutineSort": {
        "type": "string",
        "description": "Field to sort a routine listing by.",
        "enum": [
          "created",
          "updated",
          "title",
          "repository"
        ]
      },
...
      "SortOrder": {
        "type": "string",
        "description": "Sort direction for a routine listing.",
        "enum": [
          "asc",
          "desc"
        ]
      },
...
    }
  }
}